Backend dev

+++ backend/src/main/java/jp/ac/dendai/backend/Service/PostService.java
@@ -22,7 +22,7 @@ public class PostService {
         this.authService = authService;
     }
 
-    public PointDto checkPoint(Long latitude, Long longitude) {
+    public PointDto checkPoint(double latitude, double longitude) {
         // TODO
         // pointServiceのgetPointByAtPositionを呼び出し、例外がthrowされた場合は例外をthrowしServiceに送る
         // それ以外は戻り値のPointをreturn
@@ -52,19 +52,19 @@ public class PostService {
 
     public List<PostDto> getPostByUserId(String userId) {
         // TODO
-        // postRepositoryのfindByUserIdを呼び出し、例外がthrowされた場合は例外をthrowしServiceに送る
+        // postRepositoryのfindByUserIdを呼び出す
         // それ以外は戻り値のList<Post>を基にList<PostDto>を作り、return
         return null;
     }
 
     @Transactional
     public PostDto createPost(
-            String userId, Long latitude, Long longitude, String content) {
+            String userId, double latitude, double longitude, String content) {
         // TODO
         // checkUserを呼び出し、認証できなかった場合はAuthenticationFailedException例外をthrow
         // ------------ここで例外が起きたらSQLをロールバックする--------------
         // | checkPointを呼び出し、Pointが存在しない場合にはPointServiceのsaveを呼び出す
-        // | PostRepositoryのsaveを呼び出し、例外がthrowされた場合は例外をthrowしServiceに送る
+        // | PostRepositoryのsaveを呼び出す
         // ----------------------------------------------------------------
         // それ以外は戻り値のPostを基にPostDtoを作り、return
         return null;
@@ -73,7 +73,7 @@ public class PostService {
     public void deletePost(String postId, String userId) {
         // TODO
         // checkUserを呼び出し、認証できなかった場合は例外をthrow
-        // PostRepositoryのdeleteを呼び出し、例外がthrowされた場合は例外をthrowしServiceに送る
+        // PostRepositoryのdeleteを呼び出す
         // それ以外はvoidをreturn
         return;
     }
